# Faster DenseMatrix symbol iterator