This code makes the symbol iterator faster 

Apparently the recursive generators are a bottle neck.

<img width="849" height="455" alt="Sin título" src="https://github.com/user-attachments/assets/0766e5b4-7876-4fce-93b4-224f10d5983c" />


```
This

abs(genericMatrix)                           1.40 µs   ±0.17%
abs(array)                                   1.26 µs   ±0.25%
abs(numberMatrix)                            1.42 µs   ±0.15%
genericMatrix.map(abs)                       4.97 µs   ±0.18%
numberMatrix.map(abs)                        4.97 µs   ±0.16%
map(genericMatrix, abs)                      4.98 µs   ±0.20%
map(numberMatrix, abs)                       4.97 µs   ±0.15%
map(array, abs)                              4.98 µs   ±0.17%
map(array, abs.signatures.number)            2.02 µs   ±0.20%
genericMatrix.map(abs.signatures.number)     1.72 µs   ±0.24%
numberMatrix.map(abs.signatures.number)      1.69 µs   ±0.23%
genericMatrix iterate                        6.59 µs   ±0.65%

Dev

abs(genericMatrix)                           1.39 µs   ±0.19%
abs(array)                                   1.25 µs   ±0.21%
abs(numberMatrix)                            1.39 µs   ±0.17%
genericMatrix.map(abs)                       5.02 µs   ±0.15%
numberMatrix.map(abs)                        5.01 µs   ±0.16%
map(genericMatrix, abs)                      5.02 µs   ±0.16%
map(numberMatrix, abs)                       5.02 µs   ±0.16%
map(array, abs)                              4.96 µs   ±0.19%
map(array, abs.signatures.number)            2.05 µs   ±0.24%
genericMatrix.map(abs.signatures.number)     1.74 µs   ±0.54%
numberMatrix.map(abs.signatures.number)      1.71 µs   ±0.22%
genericMatrix iterate                       10.03 µs   ±1.04%
```